pyproject.toml
[project]
name = "ex2-transformer-v3-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "pillow>=11.2.1",
    "torch>=2.7.1",
    "transformers>=4.53.0",
]
